requirements.txt
langchain-astradb
langchain
langchain-community
langchain-groq
datasets
pypdf
python-dotenv
pandas
flask